% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_espn_mma_schedule.R
\name{extract_event_type_from_name}
\alias{extract_event_type_from_name}
\title{Extract event type from event name}
\usage{
extract_event_type_from_name(event_name)
}
\arguments{
\item{event_name}{Event name string}
}
\value{
Event type
}
\description{
Extract event type from event name
}
