% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_espn_mma_schedule.R
\name{extract_promotion_from_name}
\alias{extract_promotion_from_name}
\title{Extract promotion from event name}
\usage{
extract_promotion_from_name(event_name)
}
\arguments{
\item{event_name}{Event name string}
}
\value{
Promotion name
}
\description{
Extract promotion from event name
}
