% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_espn_mma_fightcenter.R
\name{get_current_featured_event}
\alias{get_current_featured_event}
\title{Get current featured event ID from ESPN MMA scoreboard}
\usage{
get_current_featured_event(league, timeout)
}
\arguments{
\item{league}{League to search for featured event}

\item{timeout}{Request timeout}
}
\value{
Event ID string or NULL if not found
}
\description{
Get current featured event ID from ESPN MMA scoreboard
}
